Update README.md
hey!! this is my first ever DS project. In this project I've applied basic learning of data cleaning. And how can we show a data through matplotlib in shape of graph. and I've used seaborn library to make graph to be look better. 
currently I'm working upon a ML model of it which can predict best value on the basis of entered data.